+++ modal/image.py
@@ -19,7 +19,6 @@ from typing import (
     Optional,
     Sequence,
     Set,
-    Tuple,
     Union,
     cast,
     get_args,
@@ -36,6 +35,7 @@ from ._utils.async_utils import synchronize_api
 from ._utils.blob_utils import MAX_OBJECT_SIZE_BYTES
 from ._utils.function_utils import FunctionInfo
 from ._utils.grpc_utils import RETRYABLE_GRPC_STATUS_CODES, retry_transient_errors
+from .client import _Client
 from .cloud_bucket_mount import _CloudBucketMount
 from .config import config, logger, user_config_path
 from .environments import _get_environment_cached
@@ -51,7 +51,6 @@ from .volume import _Volume
 
 if typing.TYPE_CHECKING:
     import modal.functions
-
 
 # This is used for both type checking and runtime validation
 ImageBuilderVersion = Literal["2023.12", "2024.04", "2024.10"]
@@ -147,8 +146,8 @@ def _get_modal_requirements_command(version: ImageBuilderVersion) -> str:
     return f"{prefix} -r {CONTAINER_REQUIREMENTS_PATH}"
 
 
-def _flatten_str_args(function_name: str, arg_name: str, args: Tuple[Union[str, List[str]], ...]) -> List[str]:
-    """Takes a tuple of strings, or string lists, and flattens it.
+def _flatten_str_args(function_name: str, arg_name: str, args: Sequence[Union[str, List[str]]]) -> List[str]:
+    """Takes a sequence of strings, or string lists, and flattens it.
 
     Raises an error if any of the elements are not strings or string lists.
     """
@@ -244,7 +243,7 @@ class _ImageRegistryConfig:
     def __init__(
         self,
         # TODO: change to _PUBLIC after worker starts handling it.
-        registry_auth_type: int = api_pb2.REGISTRY_AUTH_TYPE_UNSPECIFIED,
+        registry_auth_type: "api_pb2.RegistryAuthType.ValueType" = api_pb2.REGISTRY_AUTH_TYPE_UNSPECIFIED,
         secret: Optional[_Secret] = None,
     ):
         self.registry_auth_type = registry_auth_type
@@ -253,7 +252,7 @@ class _ImageRegistryConfig:
     def get_proto(self) -> api_pb2.ImageRegistryConfig:
         return api_pb2.ImageRegistryConfig(
             registry_auth_type=self.registry_auth_type,
-            secret_id=(self.secret.object_id if self.secret else None),
+            secret_id=(self.secret.object_id if self.secret else ""),
         )
 
 
@@ -262,6 +261,45 @@ class DockerfileSpec:
     # Ideally we would use field() with default_factory=, but doesn't work with synchronicity type-stub gen
     commands: List[str]
     context_files: Dict[str, str]
+
+
+async def _image_await_build_result(image_id: str, client: _Client) -> api_pb2.ImageJoinStreamingResponse:
+    last_entry_id: str = ""
+    result_response: Optional[api_pb2.ImageJoinStreamingResponse] = None
+
+    async def join():
+        nonlocal last_entry_id, result_response
+
+        request = api_pb2.ImageJoinStreamingRequest(image_id=image_id, timeout=55, last_entry_id=last_entry_id)
+        async for response in client.stub.ImageJoinStreaming.unary_stream(request):
+            if response.entry_id:
+                last_entry_id = response.entry_id
+            if response.result.status:
+                result_response = response
+                # can't return yet, since there may still be logs streaming back in subsequent responses
+            for task_log in response.task_logs:
+                if task_log.task_progress.pos or task_log.task_progress.len:
+                    assert task_log.task_progress.progress_type == api_pb2.IMAGE_SNAPSHOT_UPLOAD
+                    if output_mgr := _get_output_manager():
+                        output_mgr.update_snapshot_progress(image_id, task_log.task_progress)
+                elif task_log.data:
+                    if output_mgr := _get_output_manager():
+                        await output_mgr.put_log_content(task_log)
+        if output_mgr := _get_output_manager():
+            output_mgr.flush_lines()
+
+    # Handle up to n exceptions while fetching logs
+    retry_count = 0
+    while result_response is None:
+        try:
+            await join()
+        except (StreamTerminatedError, GRPCError) as exc:
+            if isinstance(exc, GRPCError) and exc.status not in RETRYABLE_GRPC_STATUS_CODES:
+                raise exc
+            retry_count += 1
+            if retry_count >= 3:
+                raise exc
+    return result_response
 
 
 class _Image(_Object, type_prefix="im"):
@@ -292,7 +330,7 @@ class _Image(_Object, type_prefix="im"):
         self._serve_mounts = other._serve_mounts
         self._deferred_mounts = other._deferred_mounts
 
-    def _hydrate_metadata(self, message: Optional[Message]):
+    def _hydrate_metadata(self, metadata: Optional[Message]):
         env_image_id = config.get("image_id")  # set as an env var in containers
         if env_image_id == self.object_id:
             for exc in self.inside_exceptions:
@@ -300,9 +338,9 @@ class _Image(_Object, type_prefix="im"):
                 # if the hydrated image is the one used by the container
                 raise exc
 
-        if message:
-            assert isinstance(message, api_pb2.ImageMetadata)
-            self._metadata = message
+        if metadata:
+            assert isinstance(metadata, api_pb2.ImageMetadata)
+            self._metadata = metadata
 
     def _add_mount_layer_or_copy(self, mount: _Mount, copy: bool = False):
         if copy:
@@ -318,7 +356,7 @@ class _Image(_Object, type_prefix="im"):
         return _Image._from_loader(_load, "Image(local files)", deps=lambda: [base_image, mount])
 
     @property
-    def _mount_layers(self) -> typing.Tuple[_Mount]:
+    def _mount_layers(self) -> typing.Sequence[_Mount]:
         """Non-evaluated mount layers on the image
 
         When the image is used by a Modal container, these mounts need to be attached as well to
@@ -344,7 +382,7 @@ class _Image(_Object, type_prefix="im"):
                 "\n"
                 "my_image = (\n"
                 "    Image.debian_slim()\n"
-                '   .add_local_python_packages("mypak", copy=True)\n'
+                '    .add_local_file("data.json", copy=True)\n'
                 '    .run_commands("python -m mypak")  # this now works!\n'
                 ")\n"
             )
@@ -362,7 +400,7 @@ class _Image(_Object, type_prefix="im"):
         context_mount: Optional[_Mount] = None,
         force_build: bool = False,
         # For internal use only.
-        _namespace: int = api_pb2.DEPLOYMENT_NAMESPACE_WORKSPACE,
+        _namespace: "api_pb2.DeploymentNamespace.ValueType" = api_pb2.DEPLOYMENT_NAMESPACE_WORKSPACE,
         _do_assert_no_mount_layers: bool = True,
     ):
         if base_images is None:
@@ -382,14 +420,14 @@ class _Image(_Object, type_prefix="im"):
         if build_function and len(base_images) != 1:
             raise InvalidError("Cannot run a build function with multiple base images!")
 
-        def _deps() -> List[_Object]:
-            deps: List[_Object] = list(base_images.values()) + list(secrets)
+        def _deps() -> Sequence[_Object]:
+            deps = tuple(base_images.values()) + tuple(secrets)
             if build_function:
-                deps.append(build_function)
+                deps += (build_function,)
             if context_mount:
-                deps.append(context_mount)
-            if image_registry_config.secret:
-                deps.append(image_registry_config.secret)
+                deps += (context_mount,)
+            if image_registry_config and image_registry_config.secret:
+                deps += (image_registry_config.secret,)
             return deps
 
         async def _load(self: _Image, resolver: Resolver, existing_object_id: Optional[str]):
@@ -398,6 +436,7 @@ class _Image(_Object, type_prefix="im"):
                     # base images can't have
                     image._assert_no_mount_layers()
 
+            assert resolver.app_id  # type narrowing
             environment = await _get_environment_cached(resolver.environment_name or "", resolver.client)
             # A bit hacky,but assume that the environment provides a valid builder version
             image_builder_version = cast(ImageBuilderVersion, environment._settings.image_builder_version)
@@ -432,7 +471,6 @@ class _Image(_Object, type_prefix="im"):
 
             if build_function:
                 build_function_id = build_function.object_id
-
                 globals = build_function._get_info().get_globals()
                 attrs = build_function._get_info().get_cls_var_attrs()
                 globals = {**globals, **attrs}
@@ -454,14 +492,14 @@ class _Image(_Object, type_prefix="im"):
 
                 # Cloudpickle function serialization produces unstable values.
                 # TODO: better way to filter out types that don't have a stable hash?
-                build_function_globals = serialize(filtered_globals) if filtered_globals else None
+                build_function_globals = serialize(filtered_globals) if filtered_globals else b""
                 _build_function = api_pb2.BuildFunction(
                     definition=build_function.get_build_def(),
                     globals=build_function_globals,
                     input=build_function_input,
                 )
             else:
-                build_function_id = None
+                build_function_id = ""
                 _build_function = None
 
             image_definition = api_pb2.Image(
@@ -470,7 +508,7 @@ class _Image(_Object, type_prefix="im"):
                 context_files=context_file_pb2s,
                 secret_ids=[secret.object_id for secret in secrets],
                 gpu=bool(gpu_config.type),  # Note: as of 2023-01-27, server still uses this
-                context_mount_id=(context_mount.object_id if context_mount else None),
+                context_mount_id=(context_mount.object_id if context_mount else ""),
                 gpu_config=gpu_config,  # Note: as of 2023-01-27, server ignores this
                 image_registry_config=image_registry_config.get_proto(),
                 runtime=config.get("function_runtime"),
@@ -481,7 +519,7 @@ class _Image(_Object, type_prefix="im"):
             req = api_pb2.ImageGetOrCreateRequest(
                 app_id=resolver.app_id,
                 image=image_definition,
-                existing_image_id=existing_object_id,  # TODO: ignored
+                existing_image_id=existing_object_id or "",  # TODO: ignored
                 build_function_id=build_function_id,
                 force_build=config.get("force_build") or force_build,
                 namespace=_namespace,
@@ -492,46 +530,22 @@ class _Image(_Object, type_prefix="im"):
             )
             resp = await retry_transient_errors(resolver.client.stub.ImageGetOrCreate, req)
             image_id = resp.image_id
+            result: api_pb2.GenericResult
+            metadata: Optional[api_pb2.ImageMetadata] = None
 
-            logger.debug("Waiting for image %s" % image_id)
-            last_entry_id: Optional[str] = None
-            result_response: Optional[api_pb2.ImageJoinStreamingResponse] = None
+            if resp.result.status:
+                # image already built
+                result = resp.result
+                if resp.HasField("metadata"):
+                    metadata = resp.metadata
+            else:
+                # not built or in the process of building - wait for build
+                logger.debug("Waiting for image %s" % image_id)
+                resp = await _image_await_build_result(image_id, resolver.client)
+                result = resp.result
+                if resp.HasField("metadata"):
+                    metadata = resp.metadata
 
-            async def join():
-                nonlocal last_entry_id, result_response
-
-                request = api_pb2.ImageJoinStreamingRequest(image_id=image_id, timeout=55, last_entry_id=last_entry_id)
-
-                async for response in resolver.client.stub.ImageJoinStreaming.unary_stream(request):
-                    if response.entry_id:
-                        last_entry_id = response.entry_id
-                    if response.result.status:
-                        result_response = response
-                        # can't return yet, since there may still be logs streaming back in subsequent responses
-                    for task_log in response.task_logs:
-                        if task_log.task_progress.pos or task_log.task_progress.len:
-                            assert task_log.task_progress.progress_type == api_pb2.IMAGE_SNAPSHOT_UPLOAD
-                            if output_mgr := _get_output_manager():
-                                output_mgr.update_snapshot_progress(image_id, task_log.task_progress)
-                        elif task_log.data:
-                            if output_mgr := _get_output_manager():
-                                await output_mgr.put_log_content(task_log)
-                if output_mgr := _get_output_manager():
-                    output_mgr.flush_lines()
-
-            # Handle up to n exceptions while fetching logs
-            retry_count = 0
-            while result_response is None:
-                try:
-                    await join()
-                except (StreamTerminatedError, GRPCError) as exc:
-                    if isinstance(exc, GRPCError) and exc.status not in RETRYABLE_GRPC_STATUS_CODES:
-                        raise exc
-                    retry_count += 1
-                    if retry_count >= 3:
-                        raise exc
-
-            result = result_response.result
             if result.status == api_pb2.GenericResult.GENERIC_STATUS_FAILURE:
                 raise RemoteError(f"Image build for {image_id} failed with the exception:\n{result.exception}")
             elif result.status == api_pb2.GenericResult.GENERIC_STATUS_TERMINATED:
@@ -545,7 +559,7 @@ class _Image(_Object, type_prefix="im"):
             else:
                 raise RemoteError("Unknown status %s!" % result.status)
 
-            self._hydrate(image_id, resolver.client, result_response.metadata)
+            self._hydrate(image_id, resolver.client, metadata)
             local_mounts = set()
             for base in base_images.values():
                 local_mounts |= base._serve_mounts
@@ -601,12 +615,59 @@ class _Image(_Object, type_prefix="im"):
             context_mount=mount,
         )
 
+    def add_local_file(self, local_path: Union[str, Path], remote_path: str, *, copy: bool = False) -> "_Image":
+        """Adds a local file to the image at `remote_path` within the container
+
+        By default (`copy=False`), the files are added to containers on startup and are not built into the actual Image,
+        which speeds up deployment.
+
+        Set `copy=True` to copy the files into an Image layer at build time instead, similar to how
+        [`COPY`](https://docs.docker.com/engine/reference/builder/#copy) works in a `Dockerfile`.
+
+        copy=True can slow down iteration since it requires a rebuild of the Image and any subsequent
+        build steps whenever the included files change, but it is required if you want to run additional
+        build steps after this one.
+        """
+        if not PurePosixPath(remote_path).is_absolute():
+            # TODO(elias): implement relative to absolute resolution using image workdir metadata
+            #  + make default remote_path="./"
+            #  This requires deferring the Mount creation until after "self" (the base image) has been resolved
+            #  so we know the workdir of the operation.
+            raise InvalidError("image.add_local_file() currently only supports absolute remote_path values")
+
+        if remote_path.endswith("/"):
+            remote_path = remote_path + Path(local_path).name
+
+        mount = _Mount.from_local_file(local_path, remote_path)
+        return self._add_mount_layer_or_copy(mount, copy=copy)
+
+    def add_local_dir(self, local_path: Union[str, Path], remote_path: str, *, copy: bool = False) -> "_Image":
+        """Adds a local directory's content to the image at `remote_path` within the container
+
+        By default (`copy=False`), the files are added to containers on startup and are not built into the actual Image,
+        which speeds up deployment.
+
+        Set `copy=True` to copy the files into an Image layer at build time instead, similar to how
+        [`COPY`](https://docs.docker.com/engine/reference/builder/#copy) works in a `Dockerfile`.
+
+        copy=True can slow down iteration since it requires a rebuild of the Image and any subsequent
+        build steps whenever the included files change, but it is required if you want to run additional
+        build steps after this one.
+        """
+        if not PurePosixPath(remote_path).is_absolute():
+            # TODO(elias): implement relative to absolute resolution using image workdir metadata
+            #  + make default remote_path="./"
+            raise InvalidError("image.add_local_dir() currently only supports absolute remote_path values")
+        mount = _Mount.from_local_dir(local_path, remote_path=remote_path)
+        return self._add_mount_layer_or_copy(mount, copy=copy)
+
     def copy_local_file(self, local_path: Union[str, Path], remote_path: Union[str, Path] = "./") -> "_Image":
         """Copy a file into the image as a part of building it.
 
         This works in a similar way to [`COPY`](https://docs.docker.com/engine/reference/builder/#copy)
         works in a `Dockerfile`.
         """
+        # TODO(elias): add pending deprecation with suggestion to use add_* instead
         basename = str(Path(local_path).name)
         mount = _Mount.from_local_file(local_path, remote_path=f"/{basename}")
 
@@ -619,7 +680,7 @@ class _Image(_Object, type_prefix="im"):
             context_mount=mount,
         )
 
-    def _add_local_python_packages(self, *packages: Union[str, Path], copy: bool = False) -> "_Image":
+    def _add_local_python_packages(self, *packages: str, copy: bool = False) -> "_Image":
         """Adds Python package files to containers
 
         Adds all files from the specified Python packages to containers running the Image.
@@ -1585,7 +1646,7 @@ class _Image(_Object, type_prefix="im"):
         function = _Function.from_args(
             info,
             app=None,
-            image=self,
+            image=self,  # type: ignore[reportArgumentType]  # TODO: probably conflict with type stub?
             secrets=secrets,
             gpu=gpu,
             mounts=mounts,
@@ -1637,7 +1698,7 @@ class _Image(_Object, type_prefix="im"):
             dockerfile_function=build_dockerfile,
         )
 
-    def workdir(self, path: str) -> "_Image":
+    def workdir(self, path: Union[str, PurePosixPath]) -> "_Image":
         """Set the working directory for subsequent image build steps and function execution.
 
         **Example**
@@ -1653,7 +1714,7 @@ class _Image(_Object, type_prefix="im"):
         """
 
         def build_dockerfile(version: ImageBuilderVersion) -> DockerfileSpec:
-            commands = ["FROM base", f"WORKDIR {shlex.quote(path)}"]
+            commands = ["FROM base", f"WORKDIR {shlex.quote(str(path))}"]
             return DockerfileSpec(commands=commands, context_files={})
 
         return _Image._from_args(
@@ -1697,7 +1758,7 @@ class _Image(_Object, type_prefix="im"):
 
         This method is considered private since its interface may change - use it at your own risk!
         """
-        last_entry_id: Optional[str] = None
+        last_entry_id: str = ""
 
         request = api_pb2.ImageJoinStreamingRequest(
             image_id=self._object_id, timeout=55, last_entry_id=last_entry_id, include_logs_for_finished=True